+++ api/filters.py
@@ -1,5 +1,5 @@
 from django_filters import rest_framework as filters
-from .models import Title
+from compositions.models import Title
 
 
 class TitlesFilter(filters.FilterSet):
