--- api_reviews&comments/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class ApiVadimConfig(AppConfig):
-    name = 'api_vadim'
