--- api_reviews&comments/pagination.py
@@ -1,12 +1,0 @@
-from rest_framework import pagination
-from rest_framework.response import Response
-
-
-class ReviewPagination(pagination.PageNumberPagination):
-    def get_paginated_response(self, data):
-        return Response({
-            'next': self.get_next_link(),
-            'previous': self.get_previous_link(),
-            'count': self.page.paginator.count,
-            'results': data
-        })
