+++ api_yamdb/urls.py
@@ -21,5 +21,4 @@ urlpatterns = [
     path('admin/', admin.site.urls),
     path('redoc/', TemplateView.as_view(template_name='redoc.html'), name='redoc'),
     path('api/', include('api.urls')),
-    path('api/', include('api_reviews&comments.urls'))
 ]
