+++ compositions/models.py
@@ -1,5 +1,24 @@
 from django.db import models
-from api.models import User, Title
+from users.models import User
+
+
+class Category(models.Model):
+    name = models.CharField(max_length=300)
+    slug = models.SlugField(unique=True)
+
+
+class Genre(models.Model):
+    name = models.CharField(max_length=300)
+    slug = models.SlugField(unique=True)
+
+
+class Title(models.Model):
+    name = models.CharField(max_length=300)
+    year = models.IntegerField(null=True, blank=True)
+    description = models.CharField(max_length=1000, blank=True)
+    rating = models.IntegerField(blank=True, null=True)
+    genre = models.ManyToManyField(Genre, verbose_name='Жанр')
+    category = models.ForeignKey(Category, on_delete=models.SET_NULL, blank=True, null=True, verbose_name='Категория')
 
 
 class Review(models.Model):
@@ -13,7 +32,7 @@ class Review(models.Model):
         (7, '7'),
         (8, '8'),
         (9, '9'),
-        (10, '10'),]
+        (10, '10'), ]
     text = models.TextField(max_length=3000)
     author = models.ForeignKey(User, on_delete=models.CASCADE, related_name='reviews')
     score = models.IntegerField(choices=RATING_CHOICES)
